test(table): split filtering test in two to optimise execution
