+++ packages/core/src/components/table/table/test/filtering/interaction.e2e.ts-snapshots/tds-table-search-clicking-on-search-button-opens-field-for-entering-data-1-linux.png
