+++ packages/core/src/components/table/table/test/filtering/interaction.e2e.ts-snapshots/tds-table-search-look-for-textbox-and-click-it-1-linux.png
